+++ packages/lighthouse-viewer/src/imported/report-assets.js
@@ -3,18 +3,23 @@
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
  * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
  */
-'use strict';
+// 'use strict';
 
-const fs = require('fs');
+// const fs = require('fs');
 
-const REPORT_TEMPLATE = fs.readFileSync(__dirname + '/assets/standalone-template.html', 'utf8');
-const REPORT_JAVASCRIPT = fs.readFileSync(__dirname + '/../dist/report/standalone.js', 'utf8');
-const REPORT_CSS = fs.readFileSync(__dirname + '/assets/styles.css', 'utf8');
+// const REPORT_TEMPLATE = fs.readFileSync(__dirname + '/assets/standalone-template.html', 'utf8');
+// const REPORT_JAVASCRIPT = fs.readFileSync(__dirname + '/../dist/report/standalone.js', 'utf8');
+// const REPORT_CSS = fs.readFileSync(__dirname + '/assets/styles.css', 'utf8');
+
+import REPORT_JAVASCRIPT from '../../dist/report/standalone.js';
+import REPORT_TEMPLATE from './assets/standalone-template.html';
+import REPORT_CSS from './assets/styles.css';
 
 // Changes to this export interface should be reflected in build/build-dt-report-resources.js
 // and clients/devtools-report-assets.js
-module.exports = {
+const assets = {
   REPORT_TEMPLATE,
   REPORT_JAVASCRIPT,
   REPORT_CSS,
-};
+}
+export default assets;