+++ packages/lighthouse-viewer/src/imported/report-generator.js
@@ -3,11 +3,12 @@
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
  * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
  */
-'use strict';
+// 'use strict';
 
-const htmlReportAssets = require('./report-assets.js');
+// const htmlReportAssets = require('./report-assets.js');
+import htmlReportAssets from './report-assets.js';
 
-class ReportGenerator {
+export class ReportGenerator {
   /**
    * Replaces all the specified strings in source without serial replacements.
    * @param {string} source
@@ -123,4 +124,4 @@ class ReportGenerator {
   }
 }
 
-module.exports = ReportGenerator;
+// module.exports = ReportGenerator;