+++ packages/lighthouse-viewer/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+
+## 1.0.0 (2021-10-27 unoffical)
+
+### Features
+
+- 'Save as HTML' generates standalone html.
+- Support changing theme. Emit event when changing theme inside lighthouse report.
+- 'Print' is removed (still avaliable in exported HTML).
+
 ## [0.1.19](https://github.com/dvelasquez/vue-lighthouse-viewer/compare/lighthouse-viewer@0.1.17...lighthouse-viewer@0.1.19) (2021-08-23)
 
 **Note:** Version bump only for package lighthouse-viewer
@@ -29,7 +38,9 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## 0.1.16 (2021-06-14)
 
-**Note:** Version bump only for package lighthouse-viewer
+**Note:** Version bump only for 
+
+ lighthouse-viewer